pyproject.toml
[project]
name = "dawn-chorus-benchmark-tools"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "faster-whisper>=1.1.0",
    "whisper-normalizer>=0.1.12",
    "jiwer>=4.0.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "datasets>=4.8.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
